server/dailycat/requirements/production.txt
Django==3.1.7
djangorestframework==3.12.2
drf-yasg==1.20.0
requests
drf-nested-routers==0.93.3
Pillow==8.2.0